feat(fe): move dummy data to dummy file
