+++ app/(main)/study/page.tsx
@@ -2,6 +2,7 @@ import SectionBanner from '@/components/common/SectionBanner'
 import { Button } from '@/components/ui/button'
 import Link from 'next/link'
 import StudyCard from '@/components/study/StudyCard'
+import { studyDummyData } from '@/lib/dummy'
 import {
   Dialog,
   DialogContent,
@@ -12,63 +13,6 @@ import {
   DialogTrigger
 } from '@/components/ui/dialog'
 
-export const studyDummyData = [
-  {
-    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
-    title: '모바일 앱 프로젝트',
-    description: '목 19:30 ~ 21:30',
-    badges: ['초급', 'Flutter', '율전']
-  },
-  {
-    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
-    title: '모바일 앱 프로젝트',
-    description: '목 19:30 ~ 21:30',
-    badges: ['초급', 'Flutter', '율전']
-  },
-  {
-    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
-    title: '모바일 앱 프로젝트',
-    description: '목 19:30 ~ 21:30',
-    badges: ['초급', 'Flutter', '율전']
-  },
-  {
-    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
-    title: '모바일 앱 프로젝트',
-    description: '목 19:30 ~ 21:30',
-    badges: ['초급', 'Flutter', '율전']
-  },
-  {
-    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
-    title: '모바일 앱 프로젝트',
-    description: '목 19:30 ~ 21:30',
-    badges: ['초급', 'Flutter', '율전']
-  },
-  {
-    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
-    title: '모바일 앱 프로젝트',
-    description: '목 19:30 ~ 21:30',
-    badges: ['초급', 'Flutter', '율전']
-  },
-  {
-    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
-    title: '모바일 앱 프로젝트',
-    description: '목 19:30 ~ 21:30',
-    badges: ['초급', 'Flutter', '율전']
-  },
-  {
-    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
-    title: '모바일 앱 프로젝트',
-    description: '목 19:30 ~ 21:30',
-    badges: [
-      '초급',
-      'Flutter',
-      '율전',
-      'Firebase',
-      '오프라인',
-      '짱 긴 text 입니다'
-    ]
-  }
-]
 export default function Study() {
   return (
     <>
