+++ lib/dummy.ts
@@ -1,0 +1,72 @@
+export const studyDummyData = [
+  {
+    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
+    title: '모바일 앱 프로젝트',
+    description: '목 19:30 ~ 21:30',
+    badges: ['초급', 'Flutter', '율전']
+  },
+  {
+    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
+    title: '모바일 앱 프로젝트',
+    description: '목 19:30 ~ 21:30',
+    badges: ['초급', 'Flutter', '율전']
+  },
+  {
+    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
+    title: '모바일 앱 프로젝트',
+    description: '목 19:30 ~ 21:30',
+    badges: ['초급', 'Flutter', '율전']
+  },
+  {
+    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
+    title: '모바일 앱 프로젝트',
+    description: '목 19:30 ~ 21:30',
+    badges: ['초급', 'Flutter', '율전']
+  },
+  {
+    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
+    title: '모바일 앱 프로젝트',
+    description: '목 19:30 ~ 21:30',
+    badges: ['초급', 'Flutter', '율전']
+  },
+  {
+    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
+    title: '모바일 앱 프로젝트',
+    description: '목 19:30 ~ 21:30',
+    badges: ['초급', 'Flutter', '율전']
+  },
+  {
+    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
+    title: '모바일 앱 프로젝트',
+    description: '목 19:30 ~ 21:30',
+    badges: ['초급', 'Flutter', '율전']
+  },
+  {
+    imageSrc: 'https://i.esdrop.com/d/f/rWQIyJ1wtg/Z2bQcFPLxZ.svg',
+    title: '모바일 앱 프로젝트',
+    description: '목 19:30 ~ 21:30',
+    badges: [
+      '초급',
+      'Flutter',
+      '율전',
+      'Firebase',
+      '오프라인',
+      '짱 긴 text 입니다'
+    ]
+  }
+]
+
+export const dummyStackUrl: string[] = [
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg',
+  'https://raw.githubusercontent.com/devicons/devicon/master/icons/python/python-original.svg'
+]
